Data.Functor.Join

Documentation

type Iso s t a b = forall p f. (Profunctor p, Functor f) => p a (f b) -> p s (f t)

An isomorphism, as in: https://hackage.haskell.org/package/lens-4.15.4/docs/Control-Lens-Iso.html

type Iso' s a = Iso s s a a

An isomorphism, specialized to two types

class (Functor f, Functor g) => Join f g where

Join two functors together

Minimal complete definition

partition, pushover, push

Methods

partition :: Iso' (f (g a)) (f (f (g a)))

pushover :: Iso' (f (f (g a))) (f (g (f a)))

push :: Iso' (f (g a)) (f (g (f a)))

type family JointL (a :: *) :: * -> * where ...

Join left

Equations

JointL (f (f (f a))) = JointL (f (f a))

JointL (f (f (g a))) = f

JointL (f (g a)) = f

JointL _ = TypeError (Text "JointL called on something other than a Join composition")

type family JointR (a :: *) :: * -> * where ...

Join right

Equations

JointR (f (f (f a))) = JointR (f (f a))

JointR (f (f (g a))) = g

JointR (f (g a)) = g

JointR _ = TypeError (Text "JointR called on something other than a Join composition")

data JoinTypes

Three types of join: left, right, and done

Constructors

JoinL

JoinR

Joined

joinL :: f (f a) -> f a

undefined

joinR :: Join f g => f (g (f a)) -> f (g a)

undefined

newtype F f a

Simple wrapper

Constructors

F

Fields

getF :: f a

newtype G g a

Simple wrapper

Constructors

G

Fields

getG :: g a

type family JoinType (a :: *) :: * where ...

The type of a join

Equations

JoinType (f (f (f a))) = F f (JoinType (f (f a)))

JoinType (f (f (g a))) = F f (JoinType (f (g a)))

JoinType (f (g (f a))) = JoinType' 1 (f (g a))

JoinType (f (g a)) = F f (G g a)

type family JoinType' (n :: Nat) (a :: *) :: * where ...

If f (g a)) then return, else recurse

Equations

JoinType' n (f (g (f a))) = JoinType' (1 + n) (f (g a))

JoinType' n (f (g a)) = JoinType'' f g n a

type family JoinType'' (f :: * -> *) (g :: * -> *) (n :: Nat) (a :: *) :: * where ...

If zero, return, else recurse

Equations

JoinType'' f g 0 a = F f (G g a)

JoinType'' f g n a = JoinType'' f g (n - 1) (f a)

class Joining' (JoinType a) (JoinType b) => Joining a b where

Next iteration of Joining

Minimal complete definition

joining

Methods

joining :: a -> b

class Joining' a b where

Generic function typeclass, helper for Joining

Minimal complete definition

joining'

Methods

joining' :: a -> b

type family JoinType''' (a :: *) :: JoinTypes where ...

Next iteration of JoinType

Equations

JoinType''' (f (f (f a))) = JoinL

JoinType''' (f (f (g a))) = JoinR

JoinType''' (f (g (f a))) = JoinL

JoinType''' (f (g a)) = Joined

class (JointL a ~ JointL b, JointR a ~ JointR b, Join (JointL a) (JointR b), JoinType''' a ~ joinTypeA, JoinType''' b ~ joinTypeB) => Joining'' a b (joinTypeA :: JoinTypes) (joinTypeB :: JoinTypes) where

For this attempt, the instances are still pretty far from compiling:

instance Joining'' a b Joined Joined => Joining'' a b JoinL JoinL where
  joining'' x = joinL . joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b JoinL JoinR where
  joining'' x = joinL . joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b JoinL Joined where
  joining'' x = joinL . joining'' x

instance Joining'' a b Joined Joined => Joining'' a b JoinR JoinL where
  joining'' x = joinR . joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b JoinR JoinR where
  joining'' x = joinR . joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b JoinR Joined where
  joining'' x = joinR . joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b Joined JoinL where
  joining'' x =         joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b Joined JoinR where
  joining'' x =         joining'' x . joinR

instance Joining'' a b Joined Joined => Joining'' a b Joined Joined where
  joining'' x = id

Minimal complete definition

joining''

Methods

joining'' :: a -> b