Data.HashSet.Utils

Documentation

type Set = HashSet

Convenience synonym

allSplits :: (Eq a, Hashable a) => Seq a -> Set (Seq a, a, Seq a)

All splits of a sequence (initial, current, rest)

prop_allSplitsLength :: [Int] -> Bool

prop_allSplitsAreSplits :: [Int] -> Bool

Each should be equal to the original, when put back together

prop_allSplitsHaveExpectedLengths :: [Int] -> Bool

allSplits_ :: (Eq a, Hashable a) => Seq a -> Set (Seq a, Seq a)

All splits of a sequence (initial, rest)

prop_allSplits_Length :: [Int] -> Bool

length seq == length set

prop_allSplits_AreSplits :: [Int] -> Bool

Each should be equal to the original, when put back together

prop_allSplits_HaveExpectedLengths :: [Int] -> Bool

flattenSet :: (Eq a, Hashable a) => Set (Set a) -> Set a

Convert to list and then flatten using the union of two sets

splitsOn :: (Eq a, Hashable a) => Seq a -> Seq a -> Set (Seq a, Seq a)

All splits of a sequence on another sequence

replacements :: (Eq a, Hashable a) => Seq a -> Set (Seq a) -> Seq a -> Set (Seq a)

All one-application replacements

quickCheckHashSetUtils :: IO Bool